+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   enabled mutually recursive functions!
 - Discarded variable names may now include numbers.
 - Fixed a bug where discarded variables might generate incorrect Erlang.
+- Added support tuple access in clause guards. 
 - New projects are created with version 1.0.2 of the setup-gleam GitHub
   action.